app/requirements.txt
langchain
langgraph
langchain-openai
langchain-neo4j
qdrant-client
neo4j-graphrag
langfuse
python-dotenv
openai
fastapi
uvicorn